Working(?) and bug fixed.
Should add a try catch exception in isPass, it will throw an error if
the image is the same with the back(no body), in this case, return fail.

Even though it is working, I have jumping up and down for 5 minutes
before I can pass one single level. It is so hard. Maybe it's just a
mask problem.

But we still have to consider DEMO hardship.

Moreover, LEFT and RIGHT is different!!! We want to add a mirro
feature.

Signed-off-by: Peter Rong <PeterRong96@gmail.com>